add dark mode
the slider I struggled with to get in a good position on the
page - it did not go nicely into the sidebar because I could
not get the top padding to work. This "on the top" seemed to be the
best option for the time being - it is not perfect because when you
go to ipad size it overlapst the logo, but someone with more css
chops can take a stab at this I hope!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>